+++ engine/src/render/camera/draws_to.cpp
@@ -1,8 +1,12 @@
 #include <cubos/core/ecs/reflection.hpp>
+#include <cubos/core/reflection/external/glm.hpp>
 
 #include <cubos/engine/render/camera/draws_to.hpp>
 
 CUBOS_REFLECT_IMPL(cubos::engine::DrawsTo)
 {
-    return core::ecs::TypeBuilder<DrawsTo>("cubos::engine::DrawsTo").build();
+    return core::ecs::TypeBuilder<DrawsTo>("cubos::engine::DrawsTo")
+        .withField("viewportOffset", &DrawsTo::viewportOffset)
+        .withField("viewportSize", &DrawsTo::viewportSize)
+        .build();
 }