+++ CHANGELOG.md
@@ -51,6 +51,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - GBuffer Rasterizer plugin (#1062, **@RiscadoA**).
 - Deferred Shading plugin (#1086, **@RiscadoA**).
 - SSAO plugin (#1088, **@RiscadoA**).
+- Split screen for the new renderer (#1149, **@tomas7770**).
 
 ### Changed